+++ src/app/OPERADORA/Servicios Iniciados/Servicios.Iniciados.Component.html
@@ -1,0 +1,73 @@
+    <!-- NAVBAR -->
+    <nav class="navbar navbar-dark bg-dark fixed-top">
+      <div class="container-fluid">
+        <a class="navbar-brand" href="#">Servicio</a>
+        <button class="navbar-toggler" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasDarkNavbar" aria-controls="offcanvasDarkNavbar">
+          <span class="navbar-toggler-icon"></span>
+        </button>
+
+        <div class="offcanvas offcanvas-end text-bg-dark" tabindex="-1" id="offcanvasDarkNavbar" aria-labelledby="offcanvasDarkNavbarLabel">
+          <div class="offcanvas-header">
+            <h5 class="offcanvas-title" id="offcanvasDarkNavbarLabel">Menu</h5>
+            <button type="button" class="btn-close btn-close-white" data-bs-dismiss="offcanvas" aria-label="Close"></button>
+          </div>
+          <div class="offcanvas-body text-center">
+            <ul class="navbar-nav justify-content-end flex-grow-1 pe-3">
+              <div class="span2">
+              <li class="nav-item">
+                <div class="d-grid gap-2">
+                <button class="btn btn-outline-success "style="font-size:20px;" aria-current="page" href="#"><i class="bi bi-person-fill-gear" ></i> Administrar Verificadores</button>
+                </div>
+              </li>
+              <br>
+              <li class="nav-item">
+                <div class="d-grid gap-2">
+                <button class="btn btn-outline-success "style="font-size:20px;" aria-current="page" href="#"><i class="bi bi-bus-front" ></i> Asignacion de Unidades</button>
+                </div>
+              </li>
+              <br>
+              <li class="nav-item">
+                <div class="d-grid gap-2">
+                <button class="btn btn-outline-success "style="font-size:20px;" aria-current="page" href="#"><i class="bi bi-clock"></i> Horarios de Servicio</button>
+                </div>
+                <br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br> <br>
+                <li class="nav-item">
+                  <button class="btn btn-outline-danger "style="font-size:20px;" aria-current="page" href="#"><i class="bi bi-box-arrow-left"></i>Cerrrar Sesion</button>
+              </li>
+              </div>
+            </ul>
+          </div>
+        </div>
+      </div>
+    </nav>
+    <!-- End NAVBAR-->
+    <br>
+    <br>
+    <br>
+<div class="container">
+  <div class="ex1">
+      <table class="table table-">
+        <thead>
+          <tr>
+            <th>Tipo Unidad</th>
+            <th>Economico</th>
+            <th>No Tarjeton</th>
+            <th>Nombre</th>
+            <th>Corrida</th>
+            <th>Accion</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr >
+            <td>Example</td>
+            <td>Example</td>
+            <td>Example</td>
+            <td>Example</td>
+            <td>Example</td>
+            <td><button class="btn btn-primary"><i class="bi bi-pencil-square"></i></button></td>
+
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  </div>
