+++ src/app/OPERADORA/Servicios Iniciados/Servicios.Iniciados.Component.scss
@@ -1,0 +1,6 @@
+div.ex1 {
+
+  height: 300px;
+  width: 100%;
+  overflow-y: auto;
+}
