+++ src/app/OPERADORA/Servicios Iniciados/Servicios.Iniciados.Component.ts
@@ -1,0 +1,13 @@
+import {Component, OnInit} from '@angular/core'
+
+@Component({
+  templateUrl:'Servicios.Iniciados.Component.html',
+  styleUrls:['Servicios.Iniciados.Component.scss']
+})
+export class ServiciosIniciadosComponent implements OnInit{
+
+
+  ngOnInit(): void {
+  }
+
+}
